Refactor MyFavoriteRecipes to Container/View Components
