+++ client/package.json
@@ -55,6 +55,7 @@
   },
   "dependencies": {
     "babel-plugin-transform-class-properties": "^6.22.0",
+    "bcrypt": "^1.0.3",
     "history": "^4.7.2",
     "jsonwebtoken": "^8.0.1",
     "moment": "^2.18.1",
